requirements.txt
apptools==4.5.0
configobj==5.0.6
cycler==0.10.0
decorator==4.4.1
envisage==4.9.0
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==3.1.2
networkx==2.4
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
pyface==6.1.2
Pygments==2.4.2
pyparsing==2.4.5
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.3.2
six==1.13.0
traits==5.2.0
traitsui==6.1.3
vtk==8.1.2
